+++ src/main/java/se/uog/course/CoursePage.java
@@ -1,8 +1,8 @@
-package se.uog.appview.pages;
+package se.uog.course;
 
 import javax.swing.JPanel;
 
-import se.uog.model.Course;
+import se.uog.application.TablePageView;
 import se.uog.table.JObjectTable;
 import se.uog.table.ObjectTableModel;
