+++ src/main/java/se/uog/course/CourseTableModel.java
@@ -1,14 +1,18 @@
-package se.uog.model;
+package se.uog.course;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.function.BiFunction;
+
 import javax.swing.DefaultListModel;
+
+import se.uog.qualification.Qualification;
 import se.uog.table.ObjectTableColumn;
 import se.uog.table.ObjectTableColumnBuilder;
 import se.uog.table.ObjectTableListSelector;
 import se.uog.table.ObjectTableModel;
+import se.uog.teacher.Teacher;
 
 public class CourseTableModel implements ObjectTableModel<Course> {
