+++ src/main/java/se/uog/database/AppModelDeserializer.java
@@ -10,12 +10,12 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 
-import se.uog.model.AppModel;
-import se.uog.model.Course;
-import se.uog.model.IDReferenced;
-import se.uog.model.Qualification;
-import se.uog.model.Teacher;
-import se.uog.model.Training;
+import se.uog.IDReferenced;
+import se.uog.application.AppModel;
+import se.uog.course.Course;
+import se.uog.qualification.Qualification;
+import se.uog.teacher.Teacher;
+import se.uog.training.Training;
 
 /**
  * A custom deserializer for the AppModel wrapper.
