+++ src/main/java/se/uog/qualification/QualificationPage.java
@@ -1,8 +1,8 @@
-package se.uog.appview.pages;
+package se.uog.qualification;
 
 import javax.swing.JPanel;
 
-import se.uog.model.Qualification;
+import se.uog.application.TablePageView;
 import se.uog.table.JObjectTable;
 import se.uog.table.ObjectTableModel;
