+++ src/main/java/se/uog/qualification/QualificationTableModel.java
@@ -1,9 +1,10 @@
-package se.uog.model;
+package se.uog.qualification;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import javax.swing.DefaultListModel;
+
 import se.uog.table.ObjectTableColumn;
 import se.uog.table.ObjectTableColumnBuilder;
 import se.uog.table.ObjectTableModel;
