+++ src/main/java/se/uog/teacher/Teacher.java
@@ -1,8 +1,13 @@
-package se.uog.model;
+package se.uog.teacher;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import se.uog.IDReferenced;
+import se.uog.qualification.Qualification;
+import se.uog.training.Training;
+
+// A teacher object has the following attributes: name, a list of qualifications, and possible training to be done. 
 public class Teacher extends IDReferenced {
     private String name;
     private List<Qualification> teacherQualifications = new ArrayList<Qualification>();
@@ -32,6 +37,10 @@ public class Teacher extends IDReferenced {
         return teacherQualifications;
     }
 
+    /**
+     * Adds a qualification to the Teacher's list of qualifications. 
+     * @param q
+     */
     public void addQualification(Qualification q) {
         teacherQualifications.add(q);
     }
