+++ src/main/java/se/uog/teacher/TeacherTableModel.java
@@ -1,8 +1,11 @@
-package se.uog.model;
+package se.uog.teacher;
 
 import java.util.ArrayList;
 import java.util.List;
+
 import javax.swing.DefaultListModel;
+
+import se.uog.qualification.Qualification;
 import se.uog.table.ObjectTableColumn;
 import se.uog.table.ObjectTableColumnBuilder;
 import se.uog.table.ObjectTableListSelector;
