+++ src/main/java/se/uog/training/Training.java
@@ -1,4 +1,7 @@
-package se.uog.model;
+package se.uog.training;
+
+import se.uog.IDReferenced;
+import se.uog.qualification.Qualification;
 
 public class Training extends IDReferenced {
     private String name;
